DjangoServer/requirements.txt
Django==2.1.3
django-extensions==2.1.4
django-redis==4.10.0
djangorestframework==3.9.0
mysqlclient==1.3.13
pytz==2018.7
redis==3.0.1
six==1.11.0
